Try async importing feeds.

+++ callbacks/message.py
@@ -28,37 +28,47 @@ def save_subscription(update, context):
     for i, feed in enumerate(feeds):
         if feed['name'] not in cached_podcasts.keys():
             try:
-                podcast = Podcast(feed['url'])
-                cached_podcasts.update({podcast.name: podcast})
+                promise = context.dispatcher.run_async(Podcast, feed_url=feed['url'])
+                podcast = promise.result()
+                if podcast:
+                    cached_podcasts.update({podcast.name: podcast})
+                    podcasts.append(podcast)
+                    subscribing_note = subscribing_note.edit_text(f"订阅中 ({len(podcasts)}/{feeds_count})")
+                else:
+                    failed_feeds.append(feed['url'])
+                    raise Exception(f"Error when adding feed {feed['url']}")
             except Exception as e: 
                 print(e)
                 failed_feeds.append(feed['url'])
                 continue
         else:
             podcast = cached_podcasts[feed['name']]
-        podcasts.append(podcast)
-        subscribing_note = subscribing_note.edit_text(f"订阅中 ({len(podcasts)}/{feeds_count})")
+            podcasts.append(podcast)
+            subscribing_note = subscribing_note.edit_text(f"订阅中 ({len(podcasts)}/{feeds_count})")
 
-    if len(podcasts):
-        user.import_feeds(podcasts)
-        newline = '\n'
-        reply = f"成功订阅 {feeds_count} 部播客！" if not len(failed_feeds) else (
-            f"成功订阅 {len(podcasts)} 部播客，部分订阅源解析失败。"
-            f"\n\n可能损坏的订阅源："
-            f"\n{newline.join(['`'+feed+'`' for feed in failed_feeds])}"
-        )
+    while len(podcasts) != len(failed_feeds) + len(podcasts):
+        pass
     else:
-            reply = "订阅失败:( \n\n请检查订阅文件以及其中的订阅源是否受损"
+        if len(podcasts):
+            user.import_feeds(podcasts)
+            newline = '\n'
+            reply = f"成功订阅 {feeds_count} 部播客！" if not len(failed_feeds) else (
+                f"成功订阅 {len(podcasts)} 部播客，部分订阅源解析失败。"
+                f"\n\n可能损坏的订阅源："
+                f"\n{newline.join(['`'+feed+'`' for feed in failed_feeds])}"
+            )
+        else:
+                reply = "订阅失败:( \n\n请检查订阅文件以及其中的订阅源是否受损"
 
-    subscribing_note.edit_text(
-        reply, 
-        reply_markup = InlineKeyboardMarkup.from_button(
-            InlineKeyboardButton(
-                "查 看 订 阅 列 表", 
-                switch_inline_query_current_chat=""
+        subscribing_note.edit_text(
+            reply, 
+            reply_markup = InlineKeyboardMarkup.from_button(
+                InlineKeyboardButton(
+                    "查 看 订 阅 列 表", 
+                    switch_inline_query_current_chat=""
+                )
             )
         )
-    )
 
 def subscribe_feed(update, context):
     context.bot.send_chat_action(chat_id = update.message.chat_id, action = 'typing')
@@ -67,34 +77,36 @@ def subscribe_feed(update, context):
         
     user = context.user_data['user']
     podcasts = context.bot_data['podcasts']
-    new_podcast = user.add_feed(feed)
-    try:
-        manage_page = ManagePage(
-            podcast_names = user.subscription.keys(), 
-            text = f"`{new_podcast.name}` 订阅成功！"
-        )
-        subscribing_message.delete()
-        message = update.message
-        message.reply_text(
-            text = manage_page.text, 
-            reply_markup=ReplyKeyboardMarkup(
-                manage_page.keyboard(), 
-                resize_keyboard=True,
-                one_time_keyboard=True
+    promise = context.dispatcher.run_async(user.add_feed, url = feed)
+    if promise.done:
+        try:
+            new_podcast = promise.result()
+            manage_page = ManagePage(
+                podcast_names = user.subscription.keys(), 
+                text = f"`{new_podcast.name}` 订阅成功！"
             )
-        )
-        podcast_page = PodcastPage(new_podcast)
-        message.reply_text(
-            text = podcast_page.text(), 
-            reply_markup=InlineKeyboardMarkup(podcast_page.keyboard())
-        )
-        message.delete()
-        new_podcast.subscribers.add(user.user_id)
-        if new_podcast.name not in podcasts.keys():
-            podcasts.update({new_podcast.name:new_podcast})
-    except Exception as e:
-        print(e)
-        subscribing_message.edit_text("订阅失败。可能是因为订阅源损坏 :(")
+            subscribing_message.delete()
+            message = update.message
+            message.reply_text(
+                text = manage_page.text, 
+                reply_markup=ReplyKeyboardMarkup(
+                    manage_page.keyboard(), 
+                    resize_keyboard=True,
+                    one_time_keyboard=True
+                )
+            )
+            podcast_page = PodcastPage(new_podcast)
+            message.reply_text(
+                text = podcast_page.text(), 
+                reply_markup=InlineKeyboardMarkup(podcast_page.keyboard())
+            )
+            message.delete()
+            new_podcast.subscribers.add(user.user_id)
+            if new_podcast.name not in podcasts.keys():
+                podcasts.update({new_podcast.name:new_podcast})
+        except Exception as e:
+            print(e)
+            subscribing_message.edit_text("订阅失败。可能是因为订阅源损坏 :(")
 
 def exit_reply_keyboard(update, context):
     message = update.message
